+++ lib/logitech_receiver/device.py
@@ -40,6 +40,7 @@ from . import settings
 from . import settings_templates
 from .common import Alert
 from .common import Battery
+from .common import _read_usb_product_string
 from .hidpp10_constants import NotificationFlag
 from .hidpp20_constants import SupportedFeature
 
@@ -67,296 +68,6 @@ class LowLevelInterface(Protocol):
 
     def close(self, handle, *args, **kwargs) -> bool:
         ...
-
-
-def _read_usb_product_string(hidraw_path):
-    """Read the USB product string from sysfs for a hidraw device path."""
-    import pathlib
-
-    try:
-        # /sys/class/hidraw/hidrawN/device/../../product → USB device product string
-        hidraw_name = pathlib.Path(hidraw_path).name
-        product_path = pathlib.Path("/sys/class/hidraw") / hidraw_name / "device" / ".." / ".." / "product"
-        product = product_path.read_text().strip()
-        return product if product else None
-    except (OSError, ValueError):
-        return None
-
-
-class CenturionReceiver:
-    """A lightweight receiver-like container for Centurion (PRO X 2 LIGHTSPEED) dongles.
-
-    Provides the Receiver interface to the UI so the dongle appears as a parent
-    with the headset as an indented child device. NOT a subclass of Receiver —
-    Receiver's __init__ does HID++ 1.0 register reads and pairing setup that
-    don't apply to Centurion.
-
-    All centurion communication (bridge, features, settings, battery) lives in
-    the child Device; this class is just a UI container + handle owner.
-    """
-
-    read_register: Callable = hidpp10.read_register
-    write_register: Callable = hidpp10.write_register
-    number = 0xFF
-    kind = None
-    isDevice = False
-    may_unpair = False
-    re_pairs = False
-    max_devices = 1
-
-    def __init__(self, low_level, handle, device_info, setting_callback=None):
-        assert handle
-        self.low_level = low_level
-        self.handle = handle
-        self.path = device_info.path
-        self.product_id = device_info.product_id
-        self.setting_callback = setting_callback
-        self.status_callback = None
-        self.notification_flags = None
-        self._devices = {}
-        self._firmware = None
-        self._dongle_features = None  # independently probed dongle features
-        self.cleanups = []
-
-        # Receiver identity
-        self.serial = None
-        self._usb_name = getattr(device_info, "product", None)
-        if not self._usb_name and self.path:
-            self._usb_name = _read_usb_product_string(self.path)
-        self.name = "Centurion Receiver"
-
-        # Dummy pairing object — lock_open stays False
-        from .receiver import Pairing
-
-        self.pairing = Pairing()
-
-        # Discover dongle features independently
-        self._discover_dongle_features()
-
-        # Read serial from dongle's CENTURION_DEVICE_INFO if available
-        if self.serial is None:
-            try:
-                s = _hidpp20.get_serial_centurion(self)
-                if s and s.strip() and s.strip().isprintable():
-                    self.serial = s.strip()
-            except Exception:
-                pass
-
-    def enable_connection_notifications(self, enable=True):
-        return False
-
-    def remaining_pairings(self, cache=True):
-        return None
-
-    def device_codename(self, n):
-        return self._usb_name
-
-    def request(self, request_id, *params, no_reply=False):
-        """Send an HID++ request directly to the dongle (not through bridge)."""
-        if self.handle:
-            return self.low_level.request(
-                self.handle, 0xFF, request_id, *params, no_reply=no_reply, long_message=True, protocol=2.0
-            )
-
-    def feature_request(self, feature, function=0x00, *params, no_reply=False):
-        """Send a feature request to the dongle using discovered feature indices."""
-        if self._dongle_features is None:
-            self._discover_dongle_features()
-        feature_int = int(feature)
-        for _feat, feat_id, index in self._dongle_features or []:
-            if feat_id == feature_int:
-                request_id = (index << 8) | (function & 0xFF)
-                return self.request(request_id, *params, no_reply=no_reply)
-        raise exceptions.FeatureNotSupported(feature)
-
-    def _discover_dongle_features(self):
-        """Independently discover features on the dongle hardware."""
-        self._dongle_features = []
-        try:
-            # Query ROOT for FEATURE_SET index
-            response = self.request(0x0000, 0x00, 0x01)
-            if response is None or response[0] == 0:
-                return
-            fs_index = response[0]
-            # Get feature count
-            count_resp = self.request(fs_index << 8)
-            if count_resp is None:
-                return
-            feature_count = count_resp[0]
-            # Enumerate features via CenturionFeatureSet (func 1 = 0x10, per-index query)
-            for idx in range(feature_count):
-                resp = self.request((fs_index << 8) | 0x10, idx)
-                if resp is None or len(resp) < 3:
-                    continue
-                feat_id = struct.unpack("!H", resp[1:3])[0]
-                try:
-                    feature = SupportedFeature(feat_id)
-                except ValueError:
-                    feature = f"unknown:{feat_id:04X}"
-                self._dongle_features.append((feature, feat_id, idx))
-            if logger.isEnabledFor(logging.DEBUG):
-                logger.debug("Centurion dongle features: %s", self._dongle_features)
-        except Exception:
-            logger.debug("Centurion dongle feature discovery failed", exc_info=True)
-
-    @property
-    def dongle_features(self):
-        """Return list of (feature, feat_id, index) tuples for dongle features."""
-        if self._dongle_features is None:
-            self._discover_dongle_features()
-        return self._dongle_features
-
-    def count(self):
-        return len([d for d in self._devices.values() if d is not None])
-
-    @property
-    def firmware(self):
-        if self._firmware is None and self.handle:
-            self._firmware = _hidpp20.get_firmware_centurion(self)
-        return self._firmware or ()
-
-    def notify_devices(self):
-        """Create child Device for the headset and trigger its initialization."""
-        # Signal receiver to UI first — tray/window need the receiver entry
-        # before a child device can be added under it.
-        self.changed(alert=Alert.NONE)
-
-        # Create child Device with receiver=self, number=1
-        pairing_info = {
-            "wpid": self.product_id,
-            "kind": None,
-            "serial": None,
-            "polling": None,
-            "power_switch": None,
-        }
-        dev = Device(
-            self.low_level,
-            self,
-            1,
-            None,
-            pairing_info=pairing_info,
-            setting_callback=self.setting_callback,
-        )
-        # Set centurion attributes on the child
-        dev.centurion = True
-        dev.product_id = self.product_id
-        dev.hidpp_long = True
-        dev._centurion_usb_name = self._usb_name
-        # Pre-set bridge index from dongle features so ping can probe the headset
-        for _feat, feat_id, idx in self._dongle_features or []:
-            if feat_id == 0x0003:  # CentPPBridge
-                dev._centurion_bridge_index = idx
-                break
-
-        self._devices[1] = dev
-        configuration.attach_to(dev)
-        dev.status_callback = self.status_callback
-
-        # Ping to determine online status.
-        # Notify UI either way — offline devices show as greyed out (matching receiver behavior).
-        online = dev.ping()
-        dev.changed(active=online)
-        if self.status_callback is not None:
-            self.status_callback(dev)
-
-    def changed(self, alert=Alert.NOTIFICATION, reason=None):
-        if self.status_callback is not None:
-            self.status_callback(self, alert=alert, reason=reason)
-
-    def status_string(self):
-        count = self.count()
-        if count == 0:
-            return "No devices."
-        return f"{count} device connected."
-
-    def close(self):
-        handle, self.handle = self.handle, None
-        for _n, d in self._devices.items():
-            if d:
-                d.close()
-        self._devices.clear()
-        for cleanup in self.cleanups:
-            cleanup(self)
-        return handle and self.low_level.close(handle)
-
-    def __del__(self):
-        self.close()
-
-    def __iter__(self):
-        for dev in self._devices.values():
-            if dev is not None:
-                yield dev
-
-    def __getitem__(self, key):
-        dev = self._devices.get(key)
-        if dev is not None:
-            return dev
-        raise IndexError(key)
-
-    def __len__(self):
-        return len([d for d in self._devices.values() if d is not None])
-
-    def __contains__(self, dev):
-        if isinstance(dev, int):
-            return self._devices.get(dev) is not None
-        return self.__contains__(dev.number)
-
-    def __bool__(self):
-        return self.handle is not None
-
-    __nonzero__ = __bool__
-
-    def __eq__(self, other):
-        return other is not None and self.kind == other.kind and self.path == other.path
-
-    def __ne__(self, other):
-        return other is None or self.kind != other.kind or self.path != other.path
-
-    def __hash__(self):
-        return self.path.__hash__()
-
-    def __str__(self):
-        return "<%s(%s,%s%s)>" % (
-            self.name.replace(" ", "") if self.name else "CenturionReceiver",
-            self.path,
-            "" if isinstance(self.handle, int) else "T",
-            self.handle,
-        )
-
-    __repr__ = __str__
-
-
-def create_centurion_receiver(low_level: LowLevelInterface, device_info, setting_callback=None):
-    """Opens a Centurion dongle and wraps it as a receiver-like container.
-
-    Creates a CenturionReceiver, discovers its features, then checks if
-    CentPPBridge (0x0003) is among them. If not, this is a direct-connected
-    device (wired headset) — close and return None so the caller can fall
-    back to create_device().
-
-    :returns: A CenturionReceiver, or None.
-    """
-    try:
-        handle = low_level.open_path(device_info.path)
-        if handle:
-            base._centurion_handles.add(int(handle))
-            cr = CenturionReceiver(low_level, handle, device_info, setting_callback)
-            # Check if any discovered feature is CentPPBridge (0x0003)
-            has_bridge = any(feat_id == 0x0003 for _, feat_id, _ in (cr.dongle_features or []))
-            if not has_bridge:
-                logger.info("Centurion device %s has no bridge, treating as direct device", device_info.path)
-                base._centurion_handles.discard(int(handle))
-                cr.handle = None  # prevent __del__ from double-closing
-                low_level.close(handle)
-                return None
-            return cr
-    except OSError as e:
-        logger.exception("open %s", device_info)
-        if e.errno == errno.EACCES:
-            raise e
-    except Exception as e:
-        logger.exception("open %s", device_info)
-        raise e
 
 
 def create_device(low_level: LowLevelInterface, device_info, setting_callback=None):
@@ -1186,3 +897,8 @@ class Device:
 
     def __del__(self):
         self.close()
+
+
+# Re-export from centurion.py — must be after Device class to avoid circular import
+from .centurion import CenturionReceiver  # noqa: E402,F401
+from .centurion import create_centurion_receiver  # noqa: E402,F401
